fix grid drawing

+++ src/net/rptools/maptool/model/SquareGrid.java
@@ -65,21 +65,21 @@ public class SquareGrid extends Grid {
         int startCol = (int)((int)(bounds.x / gridSize) * gridSize);
         int startRow = (int)((int)(bounds.y / gridSize) * gridSize);
         
-        for (int row = startRow; row < bounds.y + bounds.height + gridSize; row += gridSize) {
+        for (double row = startRow; row < bounds.y + bounds.height + gridSize; row += gridSize) {
             
             if (AppState.getGridSize() == 1) {
-                g.drawLine(bounds.x, row + offY, bounds.x+bounds.width, row + offY);
+                g.drawLine(bounds.x, (int)(row + offY), bounds.x+bounds.width, (int)(row + offY));
             } else {
-            	g.fillRect(bounds.x, row + offY - (AppState.getGridSize()/2), bounds.width, AppState.getGridSize());
+            	g.fillRect(bounds.x, (int)(row + offY - (AppState.getGridSize()/2)), bounds.width, AppState.getGridSize());
             }
         }
 
-        for (int col = startCol; col < bounds.x + bounds.width + gridSize; col += gridSize) {
+        for (double col = startCol; col < bounds.x + bounds.width + gridSize; col += gridSize) {
             
             if (AppState.getGridSize() == 1) {
-                g.drawLine(col + offX, bounds.y, col + offX, bounds.y + bounds.height);
+                g.drawLine((int)(col + offX), bounds.y, (int)(col + offX), bounds.y + bounds.height);
             } else {
-            	g.fillRect(col + offX - (AppState.getGridSize()/2), bounds.y, AppState.getGridSize(), bounds.height);
+            	g.fillRect((int)(col + offX - (AppState.getGridSize()/2)), bounds.y, AppState.getGridSize(), bounds.height);
             }
         }
 	}
